can view the messages on the index.html console and page by replacing an element on page with your items but this doesn't work on messageboard.html which is where you acc want your messages.
think is because I executes the postItem and postMessage functions on a html page and then want to show them on naother. but it only seems to show them on the same one. so can I change the console the messages appear in? Or initate the function on one page but execute hte result on another?